requirements.txt
langchain
langchain-google-genai
langchain-community
pydantic
python-dotenv
pypdf
pandas
openpyxl
tenacity
